Allow WP CLI commands to run
Every time I run WP CLI commands on a site while logged in over SSH, commands fail and the DUO plugin generates a bunch of garbage output in the shell. This PR adds a check for the `WP_CLI` constant in the same way that XMLRPC requests are handled, in the `duo_auth_enabled` function.